DATACMNS-1126 - Address review feedback for EntityInstantiator.
Refactor ClassGeneratingEntityInstantiator to ClassGeneratingEntityInstantiator without Kotlin-specifics and ClassGeneratingKotlinEntityInstantiator extending ClassGeneratingEntityInstantiator.

Apply intantiation argument caching to ClassGeneratingEntityInstantiator.
